fixing clippy issues

+++ helix-db/src/helix_engine/traversal_core/ops/source/n_from_type.rs
@@ -68,7 +68,7 @@ impl<'db, 'arena, 'txn, 's, I: Iterator<Item = Result<TraversalValue<'arena>, Gr
                     "value length is not at least the header length plus the label length meaning there has been a corruption on node insertion"
                 );
                 let label_in_lmdb = &value[LMDB_STRING_HEADER_LENGTH
-                    ..LMDB_STRING_HEADER_LENGTH + length_of_label_in_lmdb as usize];
+                    ..LMDB_STRING_HEADER_LENGTH + length_of_label_in_lmdb];
     
                 if label_in_lmdb == label_as_bytes {
                     match Node::<'arena>::from_bincode_bytes(id, value, self.arena) {
